Migrate to new package and dart version
* implemented background data messages, so the app will work while minimized
* converted list initializations to list literal
* use ScaffoldMessenger for snackbar because the other method was deprecated
* made other changes based on dart suggestion or deprecation
* Use autofill hints on login page to hopefully support password managers
* Converted android project with the new default empty way

+++ lib/models/shopping_list.dart
@@ -1,3 +1,5 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:nssl/firebase/cloud_messsaging.dart';
 import 'package:nssl/models/shopping_item.dart';
@@ -8,39 +10,63 @@ import 'package:nssl/server_communication/return_classes.dart';
 import 'package:nssl/server_communication/shopping_list_sync.dart';
 
 class ShoppingList {
+  static ShoppingList empty = ShoppingList.messaging(1, "", false);
+
   int id;
+
   String name;
-  List<ShoppingItem> shoppingItems = new List<ShoppingItem>();
+  List<ShoppingItem> shoppingItems = <ShoppingItem>[];
   bool messagingEnabled = true;
 
-  Future save() async {
-    await DatabaseManager.database.transaction((z) async {
-      int count = await z.rawUpdate('UPDATE ShoppingLists SET name = ?, messaging = ? WHERE id = ?', [name, messagingEnabled ? 1 : 0, id]);
-      if (count == 0)
-        await z.rawInsert('INSERT INTO ShoppingLists(id, name, messaging, user_id) VALUES(?, ?, ?, ?)', [id, name, messagingEnabled ? 1 : 0, User.ownId]);
+  ShoppingList(this.id, this.name);
 
-      await z.rawDelete("DELETE FROM ShoppingItems WHERE res_list_id = ?", [id]);
+  ShoppingList.messaging(this.id, this.name, this.messagingEnabled);
+
+  Future save() async {
+    if (!Platform.isAndroid) return;
+    await DatabaseManager.database.transaction((z) async {
+      await z.execute('INSERT OR REPLACE INTO ShoppingLists(id, name, messaging, user_id) VALUES(?, ?, ?, ?)',
+          [id, name, messagingEnabled ? 1 : 0, User.ownId]);
+
+      await z.rawDelete("DELETE FROM ShoppingItems WHERE id not in (?)", [shoppingItems.map((e) => e.id).join(",")]);
       for (var item in shoppingItems) {
-        await z.execute("INSERT OR REPLACE INTO ShoppingItems(id, name, amount, crossed, res_list_id, sortorder) VALUES (?, ?, ?, ?, ?, ?)",
+        await z.execute(
+            "INSERT OR REPLACE INTO ShoppingItems(id, name, amount, crossed, res_list_id, sortorder) VALUES (?, ?, ?, ?, ?, ?)",
             [item.id, item.name, item.amount, item.crossedOut ? 1 : 0, id, item.sortOrder]);
       }
     });
   }
 
+  Future addSingleItem(ShoppingItem item, {int index = -1}) async {
+    if (index < 0) index = shoppingItems.length;
+    shoppingItems.insert(index, item);
+    if (!Platform.isAndroid) return;
+    await DatabaseManager.database.execute(
+        "INSERT OR REPLACE INTO ShoppingItems(id, name, amount, crossed, res_list_id, sortorder) VALUES (?, ?, ?, ?, ?, ?)",
+        [item.id, item.name, item.amount, item.crossedOut ? 1 : 0, id, item.sortOrder]);
+  }
+
+  Future deleteSingleItem(ShoppingItem item) async {
+    shoppingItems.remove(item);
+    if (!Platform.isAndroid) return;
+    await DatabaseManager.database.rawDelete("DELETE FROM ShoppingItems WHERE id = ?", [item.id]);
+  }
+
   static Future<List<ShoppingList>> load() async {
+    if (!Platform.isAndroid) return <ShoppingList>[];
     var lists = await DatabaseManager.database.rawQuery("SELECT * FROM ShoppingLists WHERE user_id = ?", [User.ownId]);
 
     var items = await DatabaseManager.database.rawQuery("SELECT * FROM ShoppingItems ORDER BY res_list_id, sortorder");
 
     // TODO: if db ordering enough for us, or do we want to order by ourself in code?
     // return lists
-    //     .map((x) => new ShoppingList()
+    //     .map((x) => ShoppingList()
     //       ..id = x["id"]
     //       ..messagingEnabled = x["messaging"] == 0 ? false : true
     //       ..name = x["name"]
     //       ..shoppingItems = (items
     //           .where((y) => y["res_list_id"] == x["id"])
-    //           .map((y) => new ShoppingItem(y["name"])
+    //           .map((y) => ShoppingItem(y["name"])
     //             ..amount = y["amount"]
     //             ..crossedOut = y["crossed"] == 0 ? false : true
     //             ..id = y["id"]
@@ -50,13 +76,11 @@ class ShoppingList {
     //     .toList();
 
     return lists
-        .map((x) => new ShoppingList()
-          ..id = x["id"]
+        .map((x) => ShoppingList(x["id"], x["name"])
           ..messagingEnabled = x["messaging"] == 0 ? false : true
-          ..name = x["name"]
           ..shoppingItems = items
               .where((y) => y["res_list_id"] == x["id"])
-              .map((y) => new ShoppingItem(y["name"])
+              .map((y) => ShoppingItem(y["name"])
                 ..amount = y["amount"]
                 ..crossedOut = y["crossed"] == 0 ? false : true
                 ..id = y["id"]
@@ -76,17 +100,23 @@ class ShoppingList {
 //      return;
 //    }
     var newList = GetListResult.fromJson(res.body);
-    var items = (await DatabaseManager.database.rawQuery("SELECT id, crossed, sortorder FROM ShoppingItems WHERE res_list_id = ?", [id]));
+    List<Map<String, dynamic>> items;
+    if (Platform.isAndroid)
+      items = (await DatabaseManager.database
+          .rawQuery("SELECT id, crossed, sortorder FROM ShoppingItems WHERE res_list_id = ?", [id]));
+    else
+      items = <Map<String, dynamic>>[];
 
     shoppingItems.clear();
     for (var item in newList.products)
-      shoppingItems.add(new ShoppingItem(item.name)
+      shoppingItems.add(ShoppingItem(item.name)
         ..id = item.id
         ..amount = item.amount
         ..changed = item.changed
         ..created = item.created
-        ..crossedOut = (items.firstWhere((x) => x["id"] == item.id, orElse: () => {"crossed": 0})["crossed"] == 0 ? false : true)
-        ..sortOrder = (items.firstWhere((x) => x["id"] == item.id, orElse: () => {"sortorder": 0})["sortorder"]));
+        ..crossedOut =
+            (items.firstWhere((x) => x["id"] == item.id, orElse: () => {"crossed": 0})["crossed"] == 0 ? false : true)
+        ..sortOrder = item.sortOrder);
 
     shoppingItems.sort((a, b) => a.sortOrder.compareTo(b.sortOrder));
     save();
@@ -95,21 +125,29 @@ class ShoppingList {
   static Future<Null> reloadAllLists([BuildContext cont]) async {
     var result = GetListsResult.fromJson((await ShoppingListSync.getLists(cont)).body);
     User.shoppingLists.clear();
-    await DatabaseManager.database.delete("ShoppingLists", where: "user_id = ?", whereArgs: [User.ownId]);
+
+    if (Platform.isAndroid)
+      await DatabaseManager.database.delete("ShoppingLists", where: "user_id = ?", whereArgs: [User.ownId]);
     //await DatabaseManager.database.rawDelete("DELETE FROM ShoppingLists where user_id = ?", [User.ownId]);
-    var items = (await DatabaseManager.database.rawQuery("SELECT id, crossed, sortorder FROM ShoppingItems"));
+
+    List<Map<String, dynamic>> items;
+    if (Platform.isAndroid)
+      items = (await DatabaseManager.database.rawQuery("SELECT id, crossed, sortorder FROM ShoppingItems"));
+    else
+      items = <Map<String, dynamic>>[];
     for (var res in result.shoppingLists) {
-      var list = new ShoppingList()
-        ..id = res.id
-        ..name = res.name
-        ..shoppingItems = new List<ShoppingItem>();
+      var list = ShoppingList(res.id, res.name);
 
       for (var item in res.products)
-        list.shoppingItems.add(new ShoppingItem(item.name)
+        list.shoppingItems.add(ShoppingItem(item.name)
           ..id = item.id
           ..amount = item.amount
-          ..crossedOut = (items.firstWhere((x) => x["id"] == item.id, orElse: () => {"crossed": 0})["crossed"] == 0 ? false : true)
+          ..crossedOut =
+              (items.firstWhere((x) => x["id"] == item.id, orElse: () => {"crossed": 0})["crossed"] == 0 ? false : true)
           ..sortOrder = (items.firstWhere((x) => x["id"] == item.id, orElse: () => {"sortorder": 0})["sortorder"]));
+
+      if (list.shoppingItems.any((element) => element.sortOrder == null))
+        for (int i = 0; i < list.shoppingItems.length; i++) list.shoppingItems[i].sortOrder = i;
 
       list.shoppingItems.sort((a, b) => a.sortOrder.compareTo(b.sortOrder));
       User.shoppingLists.add(list);
@@ -119,7 +157,11 @@ class ShoppingList {
     }
   }
 
-  void subscribeForFirebaseMessaging() => firebaseMessaging.subscribeToTopic(id.toString() + "shoppingListTopic");
+  void subscribeForFirebaseMessaging() {
+    if (Platform.isAndroid) firebaseMessaging.subscribeToTopic(id.toString() + "shoppingListTopic");
+  }
 
-  void unsubscribeFromFirebaseMessaging() => firebaseMessaging.unsubscribeFromTopic(id.toString() + "shoppingListTopic");
+  void unsubscribeFromFirebaseMessaging() {
+    if (Platform.isAndroid) firebaseMessaging.unsubscribeFromTopic(id.toString() + "shoppingListTopic");
+  }
 }
